events: Move membership change logic into its own module
